fix: Locust tests correctly added
