+++ app/modules/explore/tests/locustfile2.py
@@ -1,90 +1,27 @@
-from flask import logging
-from locust import HttpUser, TaskSet, task
-from core.locust.common import get_csrf_token
-from core.environment.host import get_host_for_locust_testing
+from locust import HttpUser, task, between
+import random
 
 
-class FilteringBehavior(TaskSet):
-
-    def on_start(self):
-        self.login()
-
-    def login(self):
-        response = self.client.get("/login")
-        csrf_token = get_csrf_token(response)
-
-        response = self.client.post("/login", data={
-            "email": "user1@example.com",
-            "password": "1234",
-            "csrf_token": csrf_token
-        })
-        if response.status_code != 200:
-            print(f"Login failed: {response.status_code}")
+class ExploreModelBehavoiur(HttpUser):
+    wait_time = between(1, 2)
 
     @task(2)
-    def filter_by_query(self):
-        """
-        Perform a filtering operation with a query.
-        """
-        print("Performing filtering by query...")
-        response = self.client.post("/explore", json={"query": "ai"})
-        if response.status_code == 200:
-            print("Filtering by query successful.")
-        else:
-            print(f"Error filtering by query: {response.status_code}")
-
-    @task(1)
-    def filter_by_features(self):
-        """
-        Perform a filtering operation by number of features.
-        """
-        print("Performing filtering by number of features...")
-        response = self.client.post("/explore", json={"number_of_features": 50})
-        if response.status_code == 200:
-            print("Filtering by number of features successful.")
-        else:
-            print(f"Error filtering by number of features: {response.status_code}")
-
-    @task(1)
-    def filter_by_products(self):
-        """
-        Perform a filtering operation by number of products.
-        """
-        print("Performing filtering by number of products...")
-        response = self.client.post("/explore", json={"number_of_products": 84})
-        if response.status_code == 200:
-            print("Filtering by number of products successful.")
-        else:
-            print(f"Error filtering by number of products: {response.status_code}")
-
-    @task(1)
-    def combined_filter(self):
-        """
-        Perform a combined filtering operation.
-        """
-        print("Performing combined filtering...")
-        response = self.client.post(
-            "/explore", json={"query": "ml", "number_of_features": 50}
-        )
-        if response.status_code == 200:
-            print("Combined filtering successful.")
-        else:
-            print(f"Error in combined filtering: {response.status_code}")
-
-    def on_stop(self):
-        """
-        Cierra sesión al terminar la sesión simulada.
-        """
-        with self.client.get("/logout", catch_response=True) as response:
+    def view_models(self):
+        with self.client.get("/explore.explore2_models", catch_response=True) as response:
             if response.status_code == 200:
-                logging.info("Logout successful.")
+                print("Explore models page loaded successfully.")
             else:
-                logging.error(f"Logout failed: {response.status_code}")
-                response.failure("Logout error.")
+                print(f"Error loading explore models page: {response.status_code}")
+                response.failure(f"Got status code {response.status_code}")
 
-
-class FilteringUser(HttpUser):
-    tasks = [FilteringBehavior]
-    min_wait = 5000
-    max_wait = 9000
-    host = get_host_for_locust_testing()
+    @task(1)
+    def download_specific_model(self):
+        file_id = random.randint(1, 100)
+        with self.client.get(f"/file/download/{file_id}", catch_response=True) as response:
+            if response.status_code == 200:
+                print(f"Model {file_id} loaded successfully.")
+            elif response.status_code == 404:
+                print(f"Model {file_id} not found.")
+            else:
+                print(f"Error loading file {file_id}: {response.status_code}")
+                response.failure(f"Got status code {response.status_code}")
